requirements.txt
arxiv>=1.4.3
beautifulsoup4>=4.12.2
click>=8.1.3
python-dotenv>=1.0.0
requests>=2.28.2
rich>=13.4.1
sqlalchemy>=2.0.12
tqdm>=4.65.0
pydantic>=1.10.7
pytest>=7.3.1
rich-click>=1.7.0